feat(Reward): Reward의 보상 메시지를 반환하는 메서드 구현
